Renamed test package.
