+++ src/test/java/de/michab/scream/frontend/SchemeParserTest.java
@@ -3,7 +3,7 @@
  *
  * Copyright © 1998-2022 Michael G. Binz
  */
-package de.michab.scream.scanner;
+package de.michab.scream.frontend;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
